+++ packages/core/src/common/utils/index.ts
@@ -2,7 +2,6 @@ export { controller } from './controller.util';
 export { displayServerURL } from './display-server-url.util';
 export { escapeProp } from './escape-prop';
 export { escape } from './escape';
-export { getAjvInstance } from './get-ajv-instance';
 export { isInFile } from './is-in-file.util';
 export { renderError } from './render-error.util';
 export { render, renderToString } from './render.util';
