+++ packages/core/src/common/validation/validate-body.hook.ts
@@ -12,7 +12,7 @@ import {
   OpenApi,
   ServiceManager
 } from '../../core';
-import { getAjvInstance } from '../utils';
+import { getAjvInstance } from './get-ajv-instance';
 import { isFunction } from './helpers';
 
 /**
