+++ packages/core/src/common/validation/validate-cookie.hook.ts
@@ -13,7 +13,7 @@ import {
   OpenApi,
   ServiceManager,
 } from '../../core';
-import { getAjvInstance } from '../utils';
+import { getAjvInstance } from './get-ajv-instance';
 import { isFunction } from './helpers';
 
 /**
